END.
Well, I finished this course, but to be honest I'm not that satisfied as I expected.
This course contains a lot of BAD parts. As main example: Adding new features means not "COOL NEW FEATURE", it means "NEW HUGE AMOUNT OF BUGS THAT TEACHER IS NOT GOING TO FIX"

Sure, I learned some basics of the PHP, but still this is only the beggining of the learning.